update docs, separate functions into different files
