+++ src/services/users/send-password-reset-link.js
@@ -1,0 +1,21 @@
+const {  emailLib } = require("../../lib")
+const config = require("../../config")
+
+
+/**
+ * Sends a password reset link to the user
+ * @param {object} { email, url } - email of the user and password reset link
+ * @returns {Promise}
+ */
+const sendResetLink = async ({ email, url }) => {
+
+	const text = `Hi,
+	\n\nPlease click on the following link to reset your password:
+	\n${url}\n\nIf you did not request this, please ignore this email.\n`
+
+	await emailLib({ to: email,
+		 subject: `Reset your ${config("ORGANIZATION_NAME")} password`, text })
+
+}
+
+module.exports = sendResetLink
